requirements.txt
Django==1.6.5
South==0.8.4
argparse==1.2.1
django-angular==0.7.2
django-filter==0.7
djangorestframework==2.3.13
oauthlib==0.6.1
psycopg2==2.5.3
python-openid==2.2.5
python-social-auth==0.1.24
requests==2.3.0
requests-oauthlib==0.4.0
six==1.6.1
wsgiref==0.1.2
